Fix Leagacy Search test to keep connection, also refactor to avoid duplication
